+++ internal/cli/extract.go
@@ -3,7 +3,7 @@
  * file 'LICENSE.md', which is part of this source code package.
  */
 
-package cmd
+package cli
 
 import (
 	"github.com/spf13/cobra"
