+++ internal/cli/form_fill.go
@@ -3,7 +3,7 @@
  * file 'LICENSE.md', which is part of this source code package.
  */
 
-package cmd
+package cli
 
 import (
 	"errors"
@@ -12,12 +12,12 @@ import (
 	"strings"
 
 	"github.com/spf13/cobra"
-	"github.com/unidoc/unicli/pdf"
+	"github.com/unidoc/unipdf-cli/pkg/pdf"
 )
 
-const formFDFMergeCmdDesc = `Fill form fields from FDF file.
+const formFillCmdDesc = `Fill form fields from JSON file.
 
-The field values specified in the FDF file template are used to fill the form
+The field values specified in the JSON file template are used to fill the form
 fields in the input PDF files. In addition, the output file form fields can be
 flattened by using the --flatten flag. The flattening process makes the form
 fields of the output files read-only by appending the form field annotation
@@ -32,23 +32,26 @@ In addition, the filled output files can be saved to a different directory
 by using the --target-dir flag.
 The command can search for PDF files inside the subdirectories of the
 specified input directories by using the --recursive flag.
+
+The "form export" command can be used to generate the JSON form fields template
+for a PDF file.
 `
 
-var formFDFMergeCmdExample = fmt.Sprintf("%s\n%s\n%s\n%s\n%s\n%s\n",
-	fmt.Sprintf("%s form fdfmerge fields.fdf file_1.pdf file_n.pdf", appName),
-	fmt.Sprintf("%s form fdfmerge -O fields.fdf file_1.pdf file_n.pdf", appName),
-	fmt.Sprintf("%s form fdfmerge -O -r -f fields.fdf file_1.pdf file_n.pdf dir_1 dir_n", appName),
-	fmt.Sprintf("%s form fdfmerge -t out_dir fields.fdf file_1.pdf file_n.pdf dir_1 dir_n", appName),
-	fmt.Sprintf("%s form fdfmerge -t out_dir -r fields.fdf file_1.pdf file_n.pdf dir_1 dir_n", appName),
-	fmt.Sprintf("%s form fdfmerge -t out_dir -r -p pass fields.fdf file_1.pdf file_n.pdf dir_1 dir_n", appName),
+var formFillCmdExample = fmt.Sprintf("%s\n%s\n%s\n%s\n%s\n%s\n",
+	fmt.Sprintf("%s form fill fields.json file_1.pdf file_n.pdf", appName),
+	fmt.Sprintf("%s form fill -O fields.json file_1.pdf file_n.pdf", appName),
+	fmt.Sprintf("%s form fill -O -r -f fields.json file_1.pdf file_n.pdf dir_1 dir_n", appName),
+	fmt.Sprintf("%s form fill -t out_dir fields.json file_1.pdf file_n.pdf dir_1 dir_n", appName),
+	fmt.Sprintf("%s form fill -t out_dir -r fields.json file_1.pdf file_n.pdf dir_1 dir_n", appName),
+	fmt.Sprintf("%s form fill -t out_dir -r -p pass fields.json file_1.pdf file_n.pdf dir_1 dir_n", appName),
 )
 
-// formFDFMergeCmd represents the form fdfmerge command
-var formFDFMergeCmd = &cobra.Command{
-	Use:                   "fdfmerge [FLAG]... FDF_FILE INPUT_FILES...",
-	Short:                 "Fill form fields from FDF file",
-	Long:                  formFDFMergeCmdDesc,
-	Example:               formFDFMergeCmdExample,
+// formFillCmd represents the form fill command
+var formFillCmd = &cobra.Command{
+	Use:                   "fill [FLAG]... JSON_FILE INPUT_FILES...",
+	Short:                 "Fill form fields from JSON file",
+	Long:                  formFillCmdDesc,
+	Example:               formFillCmdExample,
 	DisableFlagsInUseLine: true,
 	Run: func(cmd *cobra.Command, args []string) {
 		// Parse input flags.
@@ -59,7 +62,7 @@ var formFDFMergeCmd = &cobra.Command{
 		flatten, _ := cmd.Flags().GetBool("flatten")
 
 		// Parse input parameters.
-		fdfPath := args[0]
+		jsonPath := args[0]
 
 		inputPaths, err := parseInputPaths(args[1:], recursive, pdfMatcher)
 		if err != nil {
@@ -84,7 +87,7 @@ var formFDFMergeCmd = &cobra.Command{
 			outputPath := generateOutputPath(inputPath, outputDir, "filled", overwrite)
 
 			// Fill input file form fields.
-			err := pdf.FormFillFDF(inputPath, fdfPath, outputPath, password, flatten)
+			err := pdf.FormFillJSON(inputPath, jsonPath, outputPath, password, flatten)
 			if err != nil {
 				printErr("Could not fill form fields: %s\n", err)
 			}
@@ -97,7 +100,7 @@ var formFDFMergeCmd = &cobra.Command{
 	},
 	Args: func(cmd *cobra.Command, args []string) error {
 		if len(args) < 2 {
-			return errors.New("must provide the FDF file and at least one input file")
+			return errors.New("must provide the JSON file and at least one input file")
 		}
 
 		return nil
@@ -105,11 +108,11 @@ var formFDFMergeCmd = &cobra.Command{
 }
 
 func init() {
-	formCmd.AddCommand(formFDFMergeCmd)
+	formCmd.AddCommand(formFillCmd)
 
-	formFDFMergeCmd.Flags().StringP("target-dir", "t", "", "output directory")
-	formFDFMergeCmd.Flags().BoolP("overwrite", "O", false, "overwrite input files")
-	formFDFMergeCmd.Flags().BoolP("recursive", "r", false, "search PDF files in subdirectories")
-	formFDFMergeCmd.Flags().StringP("password", "p", "", "input file password")
-	formFDFMergeCmd.Flags().BoolP("flatten", "f", false, "flatten form annotations")
+	formFillCmd.Flags().StringP("target-dir", "t", "", "output directory")
+	formFillCmd.Flags().BoolP("overwrite", "O", false, "overwrite input files")
+	formFillCmd.Flags().BoolP("recursive", "r", false, "search PDF files in subdirectories")
+	formFillCmd.Flags().StringP("password", "p", "", "input file password")
+	formFillCmd.Flags().BoolP("flatten", "f", false, "flatten form annotations")
 }
