+++ internal/cli/info.go
@@ -3,14 +3,14 @@
  * file 'LICENSE.md', which is part of this source code package.
  */
 
-package cmd
+package cli
 
 import (
 	"errors"
 	"fmt"
 
 	"github.com/spf13/cobra"
-	"github.com/unidoc/unicli/pdf"
+	"github.com/unidoc/unipdf-cli/pkg/pdf"
 )
 
 const infoCmdDesc = `Outputs information about the input file.
