+++ pkg/pdf/encrypt.go
