+++ pkg/pdf/explode.go
