+++ pkg/pdf/grayscale.go
