+++ pkg/pdf/optimize.go
