+++ pkg/pdf/search.go
